user alert in signup, relation between career and menu, order lists c…

+++ app/Http/Controllers/CareerController.php
@@ -9,6 +9,8 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use SimpleSoftwareIO\QrCode\Facades\QrCode;
+use App\Models\qr_code;
 
 class CareerController extends Controller
 {
@@ -26,17 +28,22 @@ class CareerController extends Controller
     {
         $roles = role::all();
         $user = Auth::user();
+        $path = null;
+        $bannerPath = null;
         if ($user->role[0]->title != 'admin') {
             $user->role[0] = $roles[1];
             $user->save();
         }
-
-        $name = $request->logo->getClientOriginalName();
-        $fullName = Str::uuid() . '_' . $name;
-        $path = $request->file('logo')->storeAs('files', $fullName, 'public');
-        $bannerName = $request->banner->getClientOriginalName();
-        $fullBannerName = Str::uuid() . '_' . $bannerName;
-        $bannerPath = $request->file('banner')->storeAs('files', $fullBannerName, 'public');
+        if (isset($request->logo)) {
+            $name = $request->logo->getClientOriginalName();
+            $fullName = Str::uuid() . '_' . $name;
+            $path = $request->file('logo')->storeAs('files', $fullName, 'public');
+        }
+        if (isset($request->banner)) {
+            $bannerName = $request->banner->getClientOriginalName();
+            $fullBannerName = Str::uuid() . '_' . $bannerName;
+            $bannerPath = $request->file('banner')->storeAs('files', $fullBannerName, 'public');
+        }
         $social_medias = json_encode($request->social_medias);
         $career_id = career::insertGetId([
             'title' => $request->title,
@@ -50,7 +57,21 @@ class CareerController extends Controller
             'description' => $request->description,
             'career_category_id' => $request->careerCategory,
             'banner' => $bannerPath,
+            'qr_count'=> $request->qr_count ? $request->qr_count : 0
         ]);
+        for ($i = 0; $i < $request->qr_count; $i++) {
+            $random = Str::random(10);
+            $link = "famenu.ir/qrcode/$career_id/" . $random;
+            $qr_svg = QrCode::size(100)->generate($link);
+            $fileName = 'qrcodes/' . $career_id . '_' . $random . '.svg';
+            Storage::disk('public')->put($fileName, $qr_svg);
+            qr_code::create([
+                'qr_path' => $fileName,
+                'career_id' => $career_id,
+                'slug' => 'qrcode/' . $career_id . '/' . $random,
+                'user_id'=> Auth::id()
+            ]);
+        }
         return to_route('career.careers', [Auth::user()]);
     }
 
@@ -77,8 +98,12 @@ class CareerController extends Controller
     public function update(Request $request)
     {
         $career = career::find($request->id);
+        $logoPath = null;
+        $bannerPath = null;
         if (isset($request->logo)) {
-            Storage::disk('public')->delete($career->logo);
+            if ($career->logo) {
+                Storage::disk('public')->delete($career->logo);
+            }
             $logoName = $request->logo->getClientOriginalName();
             $fullLogoName = Str::uuid() . '_' . $logoName;
             $logoPath = $request->file('logo')->storeAs('files', $fullLogoName, 'public');
@@ -101,6 +126,29 @@ class CareerController extends Controller
         $career->description = $request->description;
         $career->email = $request->email;
         $career->user_id = $request->user_id;
+        if ($request->qr_count) {
+            if ((int) $request->qr_count > $career->qr_count) {
+                $qr_count = (int) $request->qr_count - $career->qr_count;
+                while ($qr_count) {
+                    $random = Str::random(10);
+                    $link = "famenu.ir/qrcodes/$career->id/" . $random;
+                    $qr_svg = QrCode::size(100)->generate($link);
+                    $fileName = 'qrcodes/' . $career->id . '_' . $random . '.svg';
+                    Storage::disk('public')->put($fileName, $qr_svg);
+                    qr_code::create([
+                        'qr_path' => $fileName,
+                        'career_id' => $career->id,
+                        'slug' => 'qrcode/' . $career->id . '/' . $random,
+                        'user_id'=>Auth::id()
+                    ]);
+                    $qr_count--;
+                }
+            }
+            // if ($request->qr_num<=$menu->qr_num) {
+            //     # code...
+            // }
+        }
+        $career->qr_count = $request->qr_count;
         $career->save();
         $user = user::find($request->user_id);
         return to_route('career.careers', ['user' => $user]);
@@ -108,26 +156,30 @@ class CareerController extends Controller
 
     public function delete(career $career)
     {
-        if ($career->menu) {
-            foreach ($career->menu->qr_codes as $qr_code) {
+        if (count($career->qr_codes)) {
+            foreach($career->qr_codes as $qr_code){
                 $qr_code->delete();
             }
-            if (count($career->menu->menu_categories)) {
-                foreach ($career->menu->menu_categories as $category) {
-                    if (count($category->menu_items)) {
-                        foreach ($category->menu_items as $item) {
-                            if (count($item->ingredients)) {
-                                foreach ($item->ingredients as $ingredients) {
-                                    $ingredients->delete();
+        }
+        if (count($career->menus)) {
+            foreach($career->menus as $menu){
+                if (count($menu->menu_categories)) {
+                    foreach ($menu->menu_categories as $category) {
+                        if (count($category->menu_items)) {
+                            foreach ($category->menu_items as $item) {
+                                if (count($item->ingredients)) {
+                                    foreach ($item->ingredients as $ingredients) {
+                                        $ingredients->delete();
+                                    }
                                 }
+                                $item->delete();
                             }
-                            $item->delete();
                         }
+                        $category->delete();
                     }
-                    $category->delete();
                 }
+                $menu->delete();
             }
-            $career->menu->delete();
         }
         $career->delete();
         return to_route('user.profile', [Auth::user()]);
@@ -142,5 +194,14 @@ class CareerController extends Controller
     public function single(career $career)
     {
         return view('admin.careers.single', ['career' => $career]);
+    }
+
+    public function menus(career $career){
+        return view('admin.careers.menuList', ['career'=>$career]);
+    }
+
+     public function qr_codes(career $career)
+    {
+        return view('admin.careers.qrcodes', ['career' => $career]);
     }
 }
